On-the-fly diff exp and new scatterplots
prep-metacells now computes pairwise sums and means in the same way as
the pipeline. These are cached and then pairwise differences are
computed (and cached) on the fly.
This replaces the old method of reading from the markers files.
The scatterplot and gene criteria was also improved to include bands for
fold change and amounts.
Also, log() is replaced with write.log()