Fix NA facet2
for gene searches, the label and bag data are augmented by the
expression levels (as alpha,heat). But if a subset of a region is
chosen, then the non-selected clusters have an NA value for the
facet2.gg param (gene name). Hack to add gene name to all clusters, but
only alpha/heat values for selected. That way, other regions appear
gray.